added updated webscraper to backend folder
